Refactor TOTP API endpoint to match new docs
