+++ server/api/totp/activate/index.js
